Split output models into seperate files.
